bonne avancée

+++ database/migrations/2024_05_22_135100_add_entretien_and_montant_to_clients_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::table('clients', function (Blueprint $table) {
             $table->string('entretien')->nullable();
-            $table->decimal('montant', 8, 2)->nullable();
+            $table->double('montant')->nullable();
         });
     }
